add comments and switch back to uint256 for now

+++ src/libraries/Hooks.sol
@@ -15,16 +15,16 @@ library Hooks {
     using SwapFeeLibrary for uint24;
     using Hooks for IHooks;
 
-    uint160 internal constant BEFORE_INITIALIZE_FLAG = 1 << 159;
-    uint160 internal constant AFTER_INITIALIZE_FLAG = 1 << 158;
-    uint160 internal constant BEFORE_ADD_LIQUIDITY_FLAG = 1 << 157;
-    uint160 internal constant AFTER_ADD_LIQUIDITY_FLAG = 1 << 156;
-    uint160 internal constant BEFORE_REMOVE_LIQUIDITY_FLAG = 1 << 155;
-    uint160 internal constant AFTER_REMOVE_LIQUIDITY_FLAG = 1 << 154;
-    uint160 internal constant BEFORE_SWAP_FLAG = 1 << 153;
-    uint160 internal constant AFTER_SWAP_FLAG = 1 << 152;
-    uint160 internal constant BEFORE_DONATE_FLAG = 1 << 151;
-    uint160 internal constant AFTER_DONATE_FLAG = 1 << 150;
+    uint256 internal constant BEFORE_INITIALIZE_FLAG = 1 << 159;
+    uint256 internal constant AFTER_INITIALIZE_FLAG = 1 << 158;
+    uint256 internal constant BEFORE_ADD_LIQUIDITY_FLAG = 1 << 157;
+    uint256 internal constant AFTER_ADD_LIQUIDITY_FLAG = 1 << 156;
+    uint256 internal constant BEFORE_REMOVE_LIQUIDITY_FLAG = 1 << 155;
+    uint256 internal constant AFTER_REMOVE_LIQUIDITY_FLAG = 1 << 154;
+    uint256 internal constant BEFORE_SWAP_FLAG = 1 << 153;
+    uint256 internal constant AFTER_SWAP_FLAG = 1 << 152;
+    uint256 internal constant BEFORE_DONATE_FLAG = 1 << 151;
+    uint256 internal constant AFTER_DONATE_FLAG = 1 << 150;
 
     struct Permissions {
         bool beforeInitialize;
